# docs/openAPI_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Expense SaaS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Transaction",
        "description": "Create a transaction within a tenant.\n\nArgs:\n    payload: TransactionCreate schema containing tenant and account context and amounts.\n    db: Async DB session.\n    user: Current authenticated user creating the transaction.\n\nReturns:\n    The created Transaction record.\n\nRaises:\n    HTTPException 404 when the account does not exist.\n    HTTPException 403 when the user is not an active member of the tenant.",
        "operationId": "create_transaction_transactions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Transactions",
        "description": "List transactions for a tenant with optional filtering.\n\nArgs:\n    start: Optional start date to filter transactions (inclusive).\n    end: Optional end date to filter transactions (inclusive).\n    category_id: Optional category id to filter by.\n    account_id: Optional account id to filter by.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    List of Transaction records matching the filters and tenant context.\n\nRaises:\n    HTTPException 403 when the user is not a member of the requested tenant.",
        "operationId": "list_transactions_transactions_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionRead"
                  },
                  "title": "Response List Transactions Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Get Transaction",
        "description": "Retrieve a single transaction if the requester belongs to the same tenant.\n\nArgs:\n    transaction_id: UUID of the transaction.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    The Transaction record when permitted.\n\nRaises:\n    HTTPException 404 when transaction not found.\n    HTTPException 403 when user is not a member of the transaction's tenant.",
        "operationId": "get_transaction_transactions__transaction_id__get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Update Transaction",
        "description": "Update a transaction. Only the creator may modify their transaction.\n\nArgs:\n    transaction_id: UUID of the transaction to update.\n    payload: TransactionUpdate schema with optional fields.\n    db: Async DB session.\n    user: Current authenticated user.\n\nReturns:\n    The updated Transaction record.\n\nRaises:\n    HTTPException 404 when transaction not found.\n    HTTPException 403 when requester is not the creator.",
        "operationId": "update_transaction_transactions__transaction_id__patch",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Delete Transaction",
        "description": "Delete a transaction. Only the creator may delete their transaction.\n\nArgs:\n    transaction_id: UUID of the transaction to delete.\n    db: Async DB session.\n    user: Current authenticated user.\n\nRaises:\n    HTTPException 404 when transaction not found.\n    HTTPException 403 when requester is not the creator.",
        "operationId": "delete_transaction_transactions__transaction_id__delete",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CategoryKind": {
        "type": "string",
        "enum": [
          "expense",
          "income"
        ],
        "title": "CategoryKind",
        "description": "Kind of category indicating whether it represents an expense or income.\n\nUsed to classify categories and transactions as either costs (expense)\nor earnings (income)."
      },
...
      "Currency": {
        "type": "string",
        "enum": [
          "BRL",
          "USD",
          "EUR"
        ],
        "title": "Currency",
        "description": "Supported currency codes used for monetary fields across the models."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionCreate": {
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Currency"
              },
              {
                "type": "null"
              }
            ],
            "default": "BRL"
          },
          "transaction_date": {
            "type": "string",
            "format": "date",
            "title": "Transaction Date"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/CategoryKind"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionSource"
              },
              {
                "type": "null"
              }
            ],
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "amount",
          "transaction_date",
          "transaction_type"
        ],
        "title": "TransactionCreate",
        "description": "Input schema to create a transaction.\n\nArgs:\n    account_id: Account to which the transaction belongs.\n    category_id: Optional category classification.\n    amount: Monetary amount.\n    currency: Currency code (defaults to BRL).\n    transaction_date: Date when the transaction occurred.\n    transaction_type: Expense or income (CategoryKind).\n    description: Optional text describing the transaction.\n    source: Optional transaction source.\n\ntenant_id is derived from the active context (access token)."
      },
      "TransactionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "account_name": {
            "type": "string",
            "title": "Account Name"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "transaction_date": {
            "type": "string",
            "format": "date",
            "title": "Transaction Date"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/CategoryKind"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "reconciled": {
            "type": "boolean",
            "title": "Reconciled"
          },
          "source": {
            "$ref": "#/components/schemas/TransactionSource"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "account_id",
          "account_name",
          "category_id",
          "category_name",
          "amount",
          "currency",
          "transaction_date",
          "transaction_type",
          "description",
          "created_by",
          "created_at",
          "updated_at",
          "reconciled",
          "source"
        ],
        "title": "TransactionRead",
        "description": "Read schema returning transaction details.\n\nAttributes:\n    id: Transaction id.\n    tenant_id: Tenant id.\n    account_id: Account id.\n    category_id: Optional category id.\n    amount: Monetary amount.\n    currency: Currency code.\n    transaction_date: Date of transaction.\n    transaction_type: Expense or income.\n    description: Optional text description.\n    created_by: User who created it.\n    created_at: Creation timestamp.\n    updated_at: Last update timestamp.\n    reconciled: Reconciliation flag.\n    source: Transaction source."
      },
      "TransactionSource": {
        "type": "string",
        "enum": [
          "manual",
          "recurring"
        ],
        "title": "TransactionSource",
        "description": "Source of a transaction describing how it was created.\n\nMANUAL: created by a user interaction.\nRECURRING: generated by a scheduled/recurring process."
      },
      "TransactionUpdate": {
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Currency"
              },
              {
                "type": "null"
              }
            ]
          },
          "transaction_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Date"
          },
          "transaction_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "reconciled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconciled"
          }
        },
        "type": "object",
        "title": "TransactionUpdate",
        "description": "Input schema for updating transactions.\n\nArgs:\n    category_id: New category id (optional).\n    amount: New amount (optional).\n    currency: New currency (optional).\n    transaction_date: New date (optional).\n    transaction_type: New type (optional).\n    description: New description (optional).\n    reconciled: Updated reconciliation flag (optional)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}